[3895] Make TreeItem independent on the global list of selected or marked items
Now that a tree item only renders itself, we can pass it its
individual selected & marked status instead of the full lists.

Also lift up the dragStart callback which itself needs to know the
full selectedTreeItemIds, but can be defined outside of TreeItem.

Bug: https://github.com/eclipse-sirius/sirius-web/issues/3895
