+++ packages/trees/frontend/sirius-components-trees/src/treeitems/TreeItem.types.ts
@@ -21,14 +21,15 @@ export interface TreeItemProps {
   depth: number;
   expanded: string[];
   maxDepth: number;
-  onExpandedElementChange: (expanded: string[], maxDepth: number) => void;
   readOnly: boolean;
   textToHighlight: string | null;
   textToFilter: string | null;
-  markedItemIds: string[];
-  treeItemActionRender?: (props: TreeItemActionProps) => React.ReactNode;
+  marked: boolean;
+  selected: boolean;
+  onExpandedElementChange: (expanded: string[], maxDepth: number) => void;
   onTreeItemClick: (event: React.MouseEvent<HTMLDivElement, MouseEvent>, item: GQLTreeItem) => void;
-  selectedTreeItemIds: string[];
+  onDragStart: React.DragEventHandler<HTMLDivElement>;
+  treeItemActionRender?: (props: TreeItemActionProps) => React.ReactNode;
 }
 
 export interface TreeItemState {
